[Test] Adjust tests related to `toChar()`
The function was deprecated and must now be called as `toInt().toChar()`.

^KT-76136

+++ compiler/testData/codegen/box/primitiveTypes/kt16732.kt
@@ -3,7 +3,7 @@
 fun valueFromDB(value: Any): Any {
     return when (value) {
         is Char -> value
-        is Number-> value.toChar()
+        is Number-> value.toInt().toChar()
         is String -> value.single()
         else -> error("Unexpected value of type Char: $value")
     }
